Implement CSA pooling with leeway and smallest solver selection
