+++ lib/Solver/CMakeLists.txt
@@ -31,6 +31,7 @@ add_library(kleaverSolver
   Z3Builder.cpp
   Z3Solver.cpp
   Z3GlobalSolver.cpp
+  Z3PoolingSolver.cpp
 )
 
 llvm_config(kleaverSolver "${USE_LLVM_SHARED}" support)
